move font loading to FontManager
